Improved handling of large records.
grep needs to seek backward to the beginning of a matching record (or multiple
records back, in case of before-context). Until now, this could fail on
non-seekable underlying streams (stdin, non-indexed gzipped files), because
we only buffered a short fixed amount of data prior to the current location.
Now we'll dynamically buffer whatever we need to satisfy the necessary seeks.

Also, some general refactoring and cleanup.